+++ packages/ricos-common/web/src/RicosTypes.ts
@@ -46,6 +46,7 @@ export interface RicosEditorProps extends RicosProps {
   toolbarSettings?: ToolbarSettings;
   onBusyChange?: OnBusyChangeFunction;
   injectedContent?: RicosContent;
+  maxTextLength?: number;
   editorEvents?: {
     subscribe: (
       event: string,