+++ packages/ricos-content/web/src/migrateSchema/fromDraft/getTextNodes.ts
@@ -5,7 +5,7 @@ import { TO_RICOS_DECORATION_TYPE } from '../consts';
 
 import { isEmpty, merge } from 'lodash';
 import { getEntity } from './getRicosEntityData';
-import { genKey } from 'draft-js';
+import { genKey } from '../generateRandomKey';
 
 type KeyType = string | number;
 type StyleType = string;